requirements.txt
# Core dependencies
fastapi[all]==0.104.1  # FastAPI with all optional dependencies
uvicorn==0.24.0  # ASGI server
SQLAlchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
python-dotenv==1.0.0
websockets==12.0  # For WebSocket support

# OpenAI dependencies
openai==1.3.5
httpx==0.25.2

# Utils
pytz==2023.3.post1
greenlet

# Development dependencies
black==23.11.0  # Code formatting
flake8==6.1.0   # Linting